Implement new TLI processor with E-mission type ellipse capability
