+++ Orbitersdk/samples/ProjectApollo/src_saturn/LVDC.cpp
@@ -9464,8 +9464,6 @@ restartprep:
 			f			True anomaly of transfer ellipse
 			*/
 
-			fprintf(lvlog, "7-parameter update: T_RP: %f, C_3: %f, Inc: %f°, e: %f, alpha_D: %f°, f: %f°, theta_N: %f° \r\n", T_RP, C_3, Inclination*DEG, e, alpha_D*DEG, f*DEG, theta_N*DEG);
-
 			alpha_D_op = 0;
 			first_op = false;
 
@@ -9845,6 +9843,27 @@ bool LVDCSV::NavigationUpdate(VECTOR3 DCSRVEC, VECTOR3 DCSVVEC, double DCSNUPTIM
 		DLTTL[4] = NUPTIM;
 		Timer2Interrupt(true);
 		fprintf(lvlog, "Navigation update received! R %f %f %f V %f %f %f T %f \r\n", DCSRVEC.x, DCSRVEC.y, DCSRVEC.z, DCSRVEC.x, DCSRVEC.y, DCSRVEC.z, DCSNUPTIM);
+		return true;
+	}
+	return false;
+}
+
+bool LVDCSV::TLITargetingUpdate(double T_RP, double C_3, double Inclination, double theta_N, double e, double alpha_D, double f)
+{
+	if (LVDC_Timebase == 5 && T_RP > LVDC_TB_ETime + 10.0)
+	{
+		TU = true;
+		TU10 = false;
+
+		this->T_RP = T_RP;
+		this->C_3 = C_3;
+		this->Inclination = Inclination;
+		this->e = e;
+		this->alpha_D = alpha_D;
+		this->f = f;
+
+		fprintf(lvlog, "TLI targeting update received! T_RP: %f, C_3: %f, Inc: %f°, e: %f, alpha_D: %f°, f: %f°, theta_N: %f°\r\n", T_RP, C_3, Inclination*DEG, e, alpha_D*DEG, f*DEG, theta_N*DEG);
+
 		return true;
 	}
 	return false;
